* server-src/amcheck.c: Fix compilation on Solaris
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/branches/3_5@7418 a8d146d6-cc15-0410-8900-af154a0219e0

+++ server-src/amcheck.c
@@ -222,7 +222,7 @@ main(
 
     set_pname("amcheck");
     if (geteuid() != getuid()) {
-	error("amcheck must not be setuid (%d, %d)", geteuid(), getuid());
+	error("amcheck must not be setuid (%d, %d)", (int)geteuid(), (int)getuid());
     }
 
     /* drop root privileges */
